requirements.txt
beautifulsoup4==4.14.2
numpy==2.3.3
pandas==2.3.3
playwright==1.55.0
python-dotenv==1.1.1
python_dateutil==2.8.2
Requests==2.32.5
tqdm==4.67.1
unidecode==1.3.7
google-cloud-bigquery==3.12.0
google-cloud-secret-manager==2.19.1